ci: distribute alpha demo application for feature validation (#13) (#49)
Closes #13

The aim of this commit is to provide an alpha version of the application, built using sources of a specific GitHub branch, so as to test few features for validation people.
The CI/CD chain is today based on GitLab CI with our internal runners so as to keep in our side the certificates and profiles.

This commit brings:
- update of documentation (both for dev team and release process)
- Fastlane lanes and features to allow builds and uploads
- Xcode configuration to be compliant with TestFlight and to use and display builds metadata

Signed-off-by: Pierre-Yves Lapersonne <pierreyves.lapersonne@orange.com>