feat: define tokens and Swift Package architectures (#33) (#55)
The aim of this commit is to define the architecture of both the OUDS iOS Swift Package and the main objects of this library: raw tokens, semantic tokens, component tokens, themes.
It defines how things can be defined, overriden and used.

This evolution is inspired from the same issue on OUDS Android (https://github.com/Orange-OpenSource/ouds-android/issues/34).

Closes #33

Acked-by: Julien Déramond <julien.deramond@orange.com>
Tested-by: Pierre-Yves Lapersonne <pierreyves.lapersonne@orange.com>
Reviewed-by: Pierre-Yves Lapersonne <pierreyves.lapersonne@orange.com>
Signed-off-by: Pierre-Yves Lapersonne <pierreyves.lapersonne@orange.com>