ci: use new tool to define Xcode version to use, freeze to Xcode 16.0
The aim of this commit is to feeze the version of Xcode to use whatever the CI/CD chain is (internal GitLab CI we controll and GitHub Action we don't control).
The tool maanging the Xcode versions has been replaced because is sunset now (see https://github.com/xcpretty/xcode-install/blob/master/MIGRATION.md).

- Xcode 16.0 is the only version we use for CI/CD (preventing GitHub Action workflows to use for example Xcode 16.2 because available)
- `xcode-install` has been replaced by `xcodes`
- The following migration guide has been applied: https://github.com/xcpretty/xcode-install/blob/master/MIGRATION.md

Signed-off-by: Pierre-Yves Lapersonne <pierreyves.lapersonne@orange.com>