docs: introduce OUDS values in README
Add more details in README to introduce the project.

We need to show, explain and save the true values and original principles of OUDS before any gatekeeping of stakeholders. We need to preserve these values ​​before any Orange entity infiltration and political abuses.
Time to assume ODS was fucked up.