docs: add Swift DocC tutorial for OUDS iOS onboarding (#919) (#1065)
Add a Swift DocC tutorial, in Apple style, to onboard newcomers to OUDS.
The tutorial explains how to add the package, choose the libraries, instanciate a theme, inject it, use tokens and components.

Closes #919

This contribution is based a bit on work procided during Hacktoberfest 2025.
Thank you @Diksha78-bot for your contribution!

Co-authored-by: Diksha Dhanaji Dabhole <dikshadabhole78@gmail.com>
Co-authored-by: Pierre-Yves Lapersonne <pierreyves.lapersonne@orange.com>
Reviewed-by: Pierre-Yves Lapersonne <pierreyves.lapersonne@orange.com>
Signed-off-by: Pierre-Yves Lapersonne <pierreyves.lapersonne@orange.com>